Create XDED-html
HTML structure for a webpage that could interact with APIs. However, please note that HTML alone won't be sufficient to handle API requests and data processing.

Also, be aware of CORS (Cross-Origin Resource Sharing) limitations when making API requests from a web browser. Depending on the APIs you're using, you might need to consider using a server-side component to handle the API requests if CORS is an issue.